requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
Django==2.1.3
django-admin-list-filter-dropdown==1.0.2
django-admin-multiple-choice-list-filter==0.1.1
django-advanced-filters==1.1.0
django-braces==1.13.0
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-jet==1.0.8
django-js-asset==1.1.0
flake8==3.6.0
idna==2.7
jet==1.0
jet-django==0.4.1
mccabe==0.6.1
Pillow==5.3.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyMySQL==0.9.2
pytz==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
six==1.11.0
urllib3==1.24.1
